Safari-611.1.10/LayoutTests/editing/text-placeholder/caret-before-zero-width-placeholder-in-content-editable-start-of-word-expected.txt

This tests that the text insertion point is before the inserted text placeholder when it's inserted before the start of a word. This test can only be run in WebKitTestRunner or DumpRenderTree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getSelection().type is "Caret"
PASS window.getSelection().baseOffset is 'Hello'.length
PASS successfullyParsed is true

TEST COMPLETE
